Video 495 [partial unit and e2e tests] - Issues 28 - creates components for st entity view.
