+++ src/app/components/layouts/app-data-layout/st/entity-filter/entity-filter.component.spec.ts
@@ -1,6 +1,7 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { EntityFilterComponent } from './entity-filter.component';
+import { RoutingLayoutDetailType } from '../../../../../interfaces/st/routing/layout';
 
 describe('EntityFilterComponent', () => {
   let component: EntityFilterComponent;
@@ -19,5 +20,18 @@ describe('EntityFilterComponent', () => {
 
   it('should create', () => {
     expect(component).toBeTruthy();
+  });
+
+  it('sets the table filter based on the route', () => {
+    const routingLayoutDetailType: RoutingLayoutDetailType = 'entities';
+    let result = component.setTableFilterBasedOnRoute(routingLayoutDetailType);
+
+    expect(result).toBeTrue();
+  });
+
+  it('should handle menu changes', () => {
+    const result = component.menuChanged("");
+ 
+    expect(result).toBeTrue();
   });
 });
